+++ src/doomstat.h
@@ -450,22 +450,29 @@ extern int center_weapon;
 extern boolean fauxdemo;
 extern boolean casual_play;
 
-// General
+#define    CASUALPLAY(x) (casual_play ? (x) : 0)
+#define NOTCASUALPLAY(x) (casual_play ? (x) : 1)
+
+// General ------------------
+
 extern int wipe_type;
 extern int over_under;
 extern int jump_crouch;
 extern int fov;
 extern int viewheight_value;
 extern int view_bobbing_percentage;
+
 enum {
   IMPACTPITCH_OFF,
   IMPACTPITCH_FALL,
   IMPACTPITCH_DAMAGE,
   IMPACTPITCH_BOTH,
 }; extern int impact_pitch;
+
 extern int breathing;
 extern int teleporter_zoom;
 extern int death_camera;
+
 enum {
   CHASECAMMODE_OFF,
   CHASECAMMODE_BACK,
@@ -473,6 +480,7 @@ enum {
 }; extern int chasecam_mode;
 extern int chasecam_distance;
 extern int chasecam_height;
+
 extern int no_menu_tint;
 extern int no_berserk_tint;
 extern int damagecount_cap;
@@ -488,10 +496,12 @@ extern int a11y_weapon_flash;
 extern int a11y_weapon_pspr;
 extern int a11y_invul_colormap;
 
-// Weapons
+// Weapons ------------------
+
 extern int no_hor_autoaim;
 extern int switch_on_pickup;
 extern int weapon_bobbing_percentage;
+
 enum {
   BOBSTYLE_VANILLA,
   BOBSTYLE_INVVANILLA,
@@ -501,16 +511,19 @@ enum {
   BOBSTYLE_INVSMOOTH,
   BOBSTYLE_QUAKE,
 }; extern int bobbing_style;
+
 extern int weapon_inertia;
 extern int weaponsquat;
 extern int translucent_pspr;
 extern int show_berserk;
 
-// Status Bar/HUD
+// Status Bar/HUD -----------
+
 extern int alt_arms;
 extern int smarttotals;
 extern int hudcolor_ms_incomp;
 extern int hudcolor_ms_comp;
+
 typedef enum {
   TIMER_USE,
   TIMER_TELEPORT,
@@ -520,19 +533,23 @@ typedef enum {
 } eventtimer_t;
 extern int event_timers[];
 
-// Enemies
+// Enemies ------------------
+
 extern int extra_gibbing_on;
 extern int bloodier_gibbing;
 extern int zdoom_item_drops;
 
-// Messages
+// Messages -----------------
+
 extern int announce_milestones;
 
-// Key Bindings
+// Key Bindings -------------
+
 extern int zoom_fov;
 extern int fancy_teleport;
 
-// CFG only
+// CFG only -----------------
+
 extern int menu_background_darkening;
 extern int chasecam_crosshair;
 extern int sp_chat;
@@ -542,6 +559,7 @@ extern int weapon_inertia_scale_pct;
 extern int sx_fix;
 extern int blink_keys;
 extern int automap_overlay_darkening;
+
 typedef enum {
   EXGIB_FIST,
   EXGIB_CSAW,
@@ -551,22 +569,25 @@ typedef enum {
 } extragibbing_t;
 extern int extra_gibbing[];
 
-// Doom Compatibility
-extern int comp_blazing2;
-extern int comp_manualdoor;
-extern int comp_switchsource;
+// Doom Compatibility -------
+
 extern int comp_bruistarget;
-extern int comp_cgundblsnd;
+extern int comp_longautoaim;
 extern int comp_lscollision;
 extern int comp_lsamnesia;
 extern int comp_fuzzyblood;
 extern int comp_nonbleeders;
+extern int comp_iosdeath;
+extern int comp_choppers;
+
+extern int comp_blazing2;
+extern int comp_manualdoor;
+extern int comp_switchsource;
+extern int comp_cgundblsnd;
 extern int comp_cgunnersfx;
 extern int comp_flamst;
 extern int comp_deadoof;
-extern int comp_iosdeath;
 extern int comp_keypal;
-extern int comp_choppers;
 
 // [Nugget] --------------------------/