+++ src/p_user.c
@@ -761,7 +761,7 @@ void P_PlayerThink (player_t* player)
     boolean intercepts_overflow_enabled = overflow[emu_intercepts].enabled;
 
     overflow[emu_intercepts].enabled = false;
-    P_AimLineAttack(player->mo, player->mo->angle, 16*64*FRACUNIT, 0);
+    P_AimLineAttack(player->mo, player->mo->angle, 16*64*FRACUNIT * (comp_longautoaim+1), 0);
     overflow[emu_intercepts].enabled = intercepts_overflow_enabled;
 
     if (linetarget) // Give some info on the thing