+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 ## New Features
 
-None.
+- **_Double Autoaim range_** setting
 
 ## Changes
 
-None.
+- **Added Explosive Hitscan support to MDK Fist**
+- **Lowered MDK (Fist) attack's autoaim range** from 2048 to 1024 units
+- **Moved `comp_bruistarget` menu item to the top of its list**
 
 ## Bug Fixes
 
-None.
+- **Explosive Hitscan cheat not being saved in savegames**
+- **MDK Fist not being accounted for by crosshair target highlighting**